Applied review remarks
