+++ src/lib/Server/Output/ValueObjectVisitor/Author.php
@@ -15,7 +15,7 @@ use Ibexa\Contracts\Rest\Output\Visitor;
 final class Author extends ValueObjectVisitor
 {
     /**
-     * @var \Ibexa\Core\FieldType\Author\Author
+     * @param \Ibexa\Core\FieldType\Author\Author $data
      */
     public function visit(Visitor $visitor, Generator $generator, mixed $data): void
     {
