+++ src/lib/Server/Output/ValueObjectVisitor/CreatedPolicy.php
@@ -7,6 +7,7 @@
 
 namespace Ibexa\Rest\Server\Output\ValueObjectVisitor;
 
+use Ibexa\Contracts\Core\Repository\Values\User\PolicyDraft;
 use Ibexa\Contracts\Rest\Output\Generator;
 use Ibexa\Contracts\Rest\Output\Visitor;
 
@@ -15,7 +16,7 @@ use Ibexa\Contracts\Rest\Output\Visitor;
  *
  * @todo coverage add unit test
  */
-class CreatedPolicy extends Policy
+class CreatedPolicy extends AbstractPolicy
 {
     /**
      * Visit struct returned by controllers.
@@ -24,7 +25,17 @@ class CreatedPolicy extends Policy
      */
     public function visit(Visitor $visitor, Generator $generator, mixed $data): void
     {
-        parent::visit($visitor, $generator, $data->policy);
+        $generator->startObjectElement('Policy');
+        $visitor->setHeader(
+            'Content-Type',
+            $generator->getMediaType(
+            $data->policy instanceof PolicyDraft ? 'PolicyDraft' : 'Policy'
+        ),
+        );
+        $visitor->setHeader('Accept-Patch', $generator->getMediaType('PolicyUpdate'));
+        $this->visitPolicyAttributes($generator, $data->policy);
+        $generator->endObjectElement('Policy');
+
         $visitor->setHeader(
             'Location',
             $this->router->generate(
